refactor(table): update import paths for dataTableTypes to types
